# api/openapi_new.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portcullis",
    "version": "0.8.0"
  },
  "paths": {
...
    "/admin/tenants": {
      "get": {
        "tags": [
          "admin-tenants"
        ],
        "summary": "List Tenants",
        "description": "List all provisioned tenants, newest first (platform admin only).",
        "operationId": "list_tenants_admin_tenants_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TenantView"
                  },
                  "title": "Response List Tenants Admin Tenants Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-tenants"
        ],
        "summary": "Provision Tenant",
        "description": "Provision a new managed tenant (platform admin only).\n\nCreates the tenant, owner account, org membership, default roles, owner API\nkey, and platform license in one call.  The response contains the one-time\nowner access token and license key; the owner password is echoed only when\nit was generated server-side.",
        "operationId": "provision_tenant_admin_tenants_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantProvisionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProvisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/usage": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "Org Usage",
        "description": "Return the org's current-month usage and plan limits.",
        "operationId": "org_usage_v1_usage_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/usage": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "Admin Usage",
        "description": "Return current-month usage for every tenant (platform admin only).\n\nThis is the billing feed the Cloud control plane consumes: one entry per\ntenant with the summed counters and the plan's monthly cap.",
        "operationId": "admin_usage_admin_usage_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsageView"
                  },
                  "title": "Response Admin Usage Admin Usage Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LicensePlan": {
        "type": "string",
        "enum": [
          "community",
          "pro",
          "enterprise"
        ],
        "title": "LicensePlan",
        "description": "Commercial plan tiers for the platform license."
      },
      "LicenseStatus": {
        "type": "string",
        "enum": [
          "active",
          "expired",
          "revoked"
        ],
        "title": "LicenseStatus",
        "description": "Lifecycle state of a platform license.\n\n``active``   \u2014 valid and in force (not expired, not revoked).\n``expired``  \u2014 past its expires_at timestamp.\n``revoked``  \u2014 invalidated by a platform admin."
      },
...
      "LicenseView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "plan": {
            "$ref": "#/components/schemas/LicensePlan"
          },
          "status": {
            "$ref": "#/components/schemas/LicenseStatus"
          },
          "seat_limit": {
            "type": "integer",
            "title": "Seat Limit"
          },
          "server_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Limit"
          },
          "issued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Issued At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "licensee": {
            "type": "string",
            "title": "Licensee"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "plan",
          "status",
          "seat_limit",
          "server_limit",
          "issued_at",
          "expires_at",
          "revoked_at",
          "licensee",
          "created_at",
          "updated_at"
        ],
        "title": "LicenseView",
        "description": "Safe response schema for a platform license.\n\nNever exposes ``key_hash`` or the license key itself; the plaintext key is\nreturned only once at issue time via ``LicenseIssueResponse``."
      },
...
      "OrgRole": {
        "type": "string",
        "enum": [
          "org_owner",
          "org_admin",
          "developer",
          "viewer"
        ],
        "title": "OrgRole",
        "description": "Organization-level roles for RBAC.\n\nHierarchy (highest to lowest):\n- org_owner: Full org control (manage users, roles, servers, billing, delete org)\n- org_admin: Manage users, servers, policies, audit logs\n- developer: Manage servers, tools, rate limits, view audit\n- viewer: Read-only access to servers, tools, audit logs"
      },
...
      "TenantProvisionRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Slug"
          },
          "owner_email": {
            "type": "string",
            "maxLength": 320,
            "minLength": 3,
            "title": "Owner Email"
          },
          "owner_full_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Owner Full Name"
          },
          "owner_password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Password"
          },
          "plan": {
            "$ref": "#/components/schemas/LicensePlan",
            "default": "pro"
          },
          "licensee": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Licensee"
          },
          "seat_limit": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Seat Limit",
            "default": 5
          },
          "server_limit": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Limit"
          },
          "expires_in_days": {
            "type": "integer",
            "maximum": 3650.0,
            "exclusiveMinimum": 0.0,
            "title": "Expires In Days",
            "default": 365
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug",
          "owner_email",
          "owner_full_name"
        ],
        "title": "TenantProvisionRequest",
        "description": "Schema for provisioning a new managed tenant (platform admin only).\n\nThe owner account is created approved with the ``org_owner`` role, and a\nplatform license is issued for the tenant.  ``owner_password`` is optional \u2014\nwhen omitted the provisioning service generates one and returns it once."
      },
      "TenantProvisionResponse": {
        "properties": {
          "tenant": {
            "$ref": "#/components/schemas/TenantView"
          },
          "owner": {
            "$ref": "#/components/schemas/UserView"
          },
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "owner_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Password"
          },
          "license": {
            "$ref": "#/components/schemas/LicenseView"
          },
          "license_key": {
            "type": "string",
            "title": "License Key"
          }
        },
        "type": "object",
        "required": [
          "tenant",
          "owner",
          "access_token",
          "license",
          "license_key"
        ],
        "title": "TenantProvisionResponse",
        "description": "Provisioning result containing one-time credentials and the license key."
      },
      "TenantView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicensePlan"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "created_at"
        ],
        "title": "TenantView",
        "description": "Safe response schema for a tenant."
      },
...
      "UsageView": {
        "properties": {
          "requests": {
            "type": "integer",
            "title": "Requests",
            "default": 0
          },
          "tool_calls": {
            "type": "integer",
            "title": "Tool Calls",
            "default": 0
          },
          "rbac_denials": {
            "type": "integer",
            "title": "Rbac Denials",
            "default": 0
          },
          "rate_limit_rejections": {
            "type": "integer",
            "title": "Rate Limit Rejections",
            "default": 0
          },
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicensePlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "monthly_request_cap": {
            "type": "integer",
            "title": "Monthly Request Cap",
            "default": 0
          },
          "enforcement_enabled": {
            "type": "boolean",
            "title": "Enforcement Enabled",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "period_start",
          "period_end"
        ],
        "title": "UsageView",
        "description": "Usage counters plus the plan context for billing-tier display."
      },
      "UserApprovalStatus": {
        "type": "string",
        "enum": [
          "approved",
          "pending",
          "rejected"
        ],
        "title": "UserApprovalStatus",
        "description": "Lifecycle state of a user account.\n\n``approved``  \u2014 account active, can authenticate.\n``pending``   \u2014 created via an org join request, waiting for an admin.\n``rejected``  \u2014 join request denied by an admin."
      },
      "UserView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "org_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Org Name"
          },
          "intended_use": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intended Use"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "approval_status": {
            "$ref": "#/components/schemas/UserApprovalStatus"
          },
          "is_platform_admin": {
            "type": "boolean",
            "title": "Is Platform Admin",
            "default": false
          },
          "org_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrgRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "org_name",
          "intended_use",
          "is_active",
          "approval_status",
          "created_at",
          "updated_at"
        ],
        "title": "UserView",
        "description": "Safe response schema for a user \u2014 never exposes password_hash.\n\n``access_token`` is present only on the admin approval response."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}